Fixed tests so that they are also valid without a clean database. Delete methods are put in Setup-methods as there are problems with the TearDown-methods not being fired properly for purge.

+++ src/test/java/dk/opendesk/webscripts/TestUtils.java
@@ -77,6 +77,7 @@ public class TestUtils {
 
     public static Boolean deleteSite(TransactionService transactionService, SiteService siteService, String siteShortName) {
         return transactionService.getRetryingTransactionHelper().doInTransaction(() -> {
+            if(siteService.hasSite(siteShortName))
             siteService.deleteSite(siteShortName);
             return true;
         });
